[BUGFIX] Correctly load extension configuration in TYPO3 >= 9.2
